* subversion/libsvn_subr/compress.c   (svn__compress_lz4): Add an explicit cast when comparing signed and    unsigned integer values.  This cast to int is safe, as the contract    of the function only allows blocks <= LZ4_MAX_INPUT_SIZE.
git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1801945 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/libsvn_subr/compress.c
@@ -307,7 +307,7 @@ svn__compress_lz4(const void *data, apr_size_t len,
   if (!compressed_data_len)
     return svn_error_create(SVN_ERR_LZ4_COMPRESSION_FAILED, NULL, NULL);
 
-  if (compressed_data_len >= len)
+  if (compressed_data_len >= (int)len)
     {
       /* Compression didn't help :(, just append the original text */
       svn_stringbuf_appendbytes(out, data, len);
